app/requirements.txt
# =============================================================================
# DeepSeek PDF OCR - Requirements
# =============================================================================
# Note: PyTorch and flash-attn are installed separately in setup.sh
# to ensure correct CUDA versions

# -----------------------------------------------------------------------------
# DeepSeek-OCR Dependencies
# -----------------------------------------------------------------------------
transformers>=4.51.1
accelerate>=0.26.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
sentencepiece>=0.1.99
tiktoken>=0.5.0

# -----------------------------------------------------------------------------
# PDF Processing
# -----------------------------------------------------------------------------
pdf2image>=1.16.0          # PDF to images (uses poppler)
PyMuPDF>=1.23.0            # Fast PDF manipulation (fitz)
pdfplumber>=0.10.0         # Extract text/tables from native PDFs
Pillow>=10.0.0             # Image processing

# -----------------------------------------------------------------------------
# API Server
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6    # File uploads
pydantic>=2.0.0
pydantic-settings>=2.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
numpy>=1.24.0
tqdm>=4.65.0               # Progress bars
python-dotenv>=1.0.0       # Environment variables
aiofiles>=23.0.0           # Async file operations

# -----------------------------------------------------------------------------
# Markdown/HTML Processing
# -----------------------------------------------------------------------------
markdown>=3.5.0            # Markdown to HTML
beautifulsoup4>=4.12.0     # HTML parsing/manipulation
lxml>=5.0.0                # Fast HTML/XML processing
html5lib>=1.1              # HTML5 parsing

# -----------------------------------------------------------------------------
# Optional: Quality & Testing
# -----------------------------------------------------------------------------
# pytest>=7.4.0
# httpx>=0.26.0            # Async HTTP client for testing
